changed atoi to atol where necessary... woops\!

+++ gcmtool/src/GCMDiskHeaderToolMain.c
@@ -257,7 +257,7 @@ int main(int argc, char **argv) {
 		} else if (CHECK_ARG(ARG_UNKNOWN1)) {
 			// they want to change the unknown1 (?)
 
-			newUnknown1 = atoi(GET_NEXT_ARG);
+			newUnknown1 = atol(GET_NEXT_ARG);
 			modUnknown1++;
 			
 		} else if (CHECK_ARG(ARG_GAME_NAME)) {
@@ -272,55 +272,55 @@ int main(int argc, char **argv) {
 		} else if (CHECK_ARG(ARG_DEBUG_OFFSET)) {
 			// they want to change the debug monitor offset...
 
-			newDebugMonitorOffset = atoi(GET_NEXT_ARG);
+			newDebugMonitorOffset = atol(GET_NEXT_ARG);
 			modDebugMonitorOffset++;
 			
 		} else if (CHECK_ARG(ARG_DEBUG_ADDRESS)) {
 			// change the debug monitor length...
 
-			newDebugMonitorAddress = atoi(GET_NEXT_ARG);
+			newDebugMonitorAddress = atol(GET_NEXT_ARG);
 			modDebugMonitorAddress++;
 			
 		} else if (CHECK_ARG(ARG_DOL_OFFSET)) {
 			//change DOL offset
 
-			newDolOffset = atoi(GET_NEXT_ARG);
+			newDolOffset = atol(GET_NEXT_ARG);
 			modDolOffset++;
 			
 		} else if (CHECK_ARG(ARG_FST_OFFSET)) {
 			// change FST offset
 			
-			newFstOffset = atoi(GET_NEXT_ARG);
+			newFstOffset = atol(GET_NEXT_ARG);
 			modFstOffset++;
 			
 		} else if (CHECK_ARG(ARG_FST_SIZE)) {
 			// change FST size
 			
-			newFstSize = atoi(GET_NEXT_ARG);
+			newFstSize = atol(GET_NEXT_ARG);
 			modFstSize++;
 			
 		} else if (CHECK_ARG(ARG_FST_SIZE_MAX)) {
 			// change FST max size (for multi-disk)
 			
-			newFstSizeMax = atoi(GET_NEXT_ARG);
+			newFstSizeMax = atol(GET_NEXT_ARG);
 			modFstSizeMax++;
 			
 		} else if (CHECK_ARG(ARG_USER_POS)) {
 			// change user position
 			
-			newUserPosition = atoi(GET_NEXT_ARG);
+			newUserPosition = atol(GET_NEXT_ARG);
 			modUserPosition++;
 
 		} else if (CHECK_ARG(ARG_USER_LEN)) {
 			// change user address
 			
-			newUserLength = atoi(GET_NEXT_ARG);
+			newUserLength = atol(GET_NEXT_ARG);
 			modUserLength++;
 			
 		} else if (CHECK_ARG(ARG_UNKNOWN2)) {
 			// change unknown
 			
-			newUnknown2 = atoi(GET_NEXT_ARG);
+			newUnknown2 = atol(GET_NEXT_ARG);
 			modUnknown2++;
 			
 		} else {
